changing files back to normal to i can do a pull request, decided to keep it the same as the functions ive added can do all games on the site too, will update examples for the ones i am interested in
