+++ examples/csgo-vods.js
@@ -1,0 +1,5 @@
+var Gosu = require("../gosuapi.js");
+Gosu.fetchVodUrls('csgo' , function(err,urls){
+	console.log(urls);
+
+});
